+++ examples/pioneer/tsconfig.json
@@ -12,5 +12,5 @@
       "quasar"
     ]
   },
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules","src"]
 }
