+++ examples/sandbox/index.html
@@ -82,6 +82,8 @@
     <button id='trezor'>Pair Trezor</button>
     <button id='ledger'>Pair Ledger</button>
     <button id='portis'>Pair Portis</button>
+    <button id='pioneer'>Pair Pioneer</button>
+
 
     <select id='keyring' style="height: 100px" size="4"></select>
   </div>
